+++ Realm/ObjectServerTests/RLMObjectServerTests.mm
@@ -1510,22 +1510,61 @@
     XCTAssertNil(error);
 }
 
+- (void)testAutomaticSyncConfiguration {
+    NSURL *server = [RLMObjectServerTests authServerURL];
+
+    // Automatic configuration should throw when there are no logged-in users.
+    XCTAssertThrows([RLMSyncConfiguration automaticConfiguration]);
+
+    RLMSyncCredentials *credsA = [RLMObjectServerTests basicCredentialsWithName:@"a" register:YES];
+    RLMSyncUser *userA = [self logInUserForCredentials:credsA server:server];
+
+    // Now that there's a logged-in user, we should be able to retrieve the configuration.
+    RLMRealmConfiguration *configuration = [RLMSyncConfiguration automaticConfiguration];
+    XCTAssert(configuration);
+
+    @autoreleasepool {
+        // And open it successfully.
+        RLMRealm *realm = [self openRealmWithConfiguration:configuration];
+        [self waitForDownloadsForRealm:realm];
+    }
+
+
+    RLMSyncCredentials *credsB = [RLMObjectServerTests basicCredentialsWithName:@"b" register:YES];
+    RLMSyncUser *userB = [self logInUserForCredentials:credsB server:server];
+
+    // Automatic configuration should throw since there's more than one logged-in user.
+    XCTAssertThrows([RLMSyncConfiguration automaticConfiguration]);
+
+    // It should still be possible to explicitly retrieve an automatic configuration for a user.
+    RLMRealmConfiguration *configurationA = [RLMSyncConfiguration automaticConfigurationForUser:userA];
+    XCTAssert(configurationA);
+    XCTAssertEqualObjects(configuration.syncConfiguration, configurationA.syncConfiguration);
+
+    RLMRealmConfiguration *configurationB = [RLMSyncConfiguration automaticConfigurationForUser:userB];
+    XCTAssert(configurationB);
+    XCTAssertNotEqualObjects(configuration.syncConfiguration, configurationB.syncConfiguration);
+
+
+    [userB logOut];
+
+    // Now that we're back to a single logged-in user, we should be able to retrieve the configuration.
+    configuration = [RLMSyncConfiguration automaticConfiguration];
+    XCTAssert(configuration);
+}
+
 #pragma mark - Partial sync
 
 - (void)testPartialSync {
     // Make credentials.
     NSString *name = NSStringFromSelector(_cmd);
     NSURL *server = [RLMObjectServerTests authServerURL];
-    NSURL *realmURL = REALM_URL();
 
     // Log in and populate the Realm.
     @autoreleasepool {
         RLMSyncCredentials *creds = [RLMObjectServerTests basicCredentialsWithName:name register:YES];
         RLMSyncUser *user = [self logInUserForCredentials:creds server:server];
-        RLMSyncConfiguration *syncConfig = [[RLMSyncConfiguration alloc] initWithUser:user realmURL:realmURL];
-        syncConfig.isPartial = YES;
-        RLMRealmConfiguration *configuration = [RLMRealmConfiguration defaultConfiguration];
-        configuration.syncConfiguration = syncConfig;
+        RLMRealmConfiguration *configuration = [RLMSyncConfiguration automaticConfigurationForUser:user];
         RLMRealm *realm = [self openRealmWithConfiguration:configuration];
         [realm beginWriteTransaction];
         // FIXME: make this less hideous
@@ -1558,10 +1597,7 @@
     @autoreleasepool {
         RLMSyncCredentials *creds = [RLMObjectServerTests basicCredentialsWithName:name register:NO];
         RLMSyncUser *user = [self logInUserForCredentials:creds server:server];
-        RLMSyncConfiguration *syncConfig = [[RLMSyncConfiguration alloc] initWithUser:user realmURL:realmURL];
-        syncConfig.isPartial = YES;
-        RLMRealmConfiguration *configuration = [RLMRealmConfiguration defaultConfiguration];
-        configuration.syncConfiguration = syncConfig;
+        RLMRealmConfiguration *configuration = [RLMSyncConfiguration automaticConfigurationForUser:user];
         RLMRealm *realm = [self openRealmWithConfiguration:configuration];
 
         // Perform some partial sync queries