+++ Realm/RLMSyncUser_Private.hpp
@@ -62,8 +62,8 @@ private:
 };
 
 @interface RLMSyncUser ()
-
 - (instancetype)initWithSyncUser:(std::shared_ptr<SyncUser>)user;
+- (NSURL *)defaultRealmURL;
 - (std::shared_ptr<SyncUser>)_syncUser;
 - (nullable NSString *)_refreshToken;
 + (void)_setUpBindingContextFactory;